5/5/2024

+++ src/main/java/com/ds/darknesschat/user/User.java
@@ -4,7 +4,7 @@ public class User {
     private final String userName;
     private final String userPassword;
     private final String userDateOfRegistration;
-    private int id = -1;
+    private long id = -1L;
 
     public User(String userName, String userPassword, String userDateOfRegistration) {
         this.userName = userName;
@@ -12,14 +12,14 @@ public class User {
         this.userDateOfRegistration = userDateOfRegistration;
     }
 
-    public User(String userName, String userPassword, String userDateOfRegistration, int id) {
+    public User(String userName, String userPassword, String userDateOfRegistration, long id) {
         this.userName = userName;
         this.userPassword = userPassword;
         this.userDateOfRegistration = userDateOfRegistration;
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
